feat(chess): update old logic with new domains
- second domain structure refactoring complete
- remove unused tests
- match test packages to main
