+++ src/main/java/chess/Application.java
@@ -4,7 +4,7 @@ import static chess.view.InputView.requestStartOrEndInput;
 import static chess.view.OutputView.printBoard;
 import static chess.view.OutputView.printGameStartAnnouncement;
 
-import chess.domain.board.Board;
+import chess.domain.ChessGame;
 import chess.dto.BoardDto;
 
 public class Application {
@@ -12,8 +12,8 @@ public class Application {
     public static void main(String[] args) {
         printGameStartAnnouncement();
         while (requestStartOrEndInput()) {
-            Board board = new Board();
-            BoardDto dto = new BoardDto(board);
+            ChessGame game = new ChessGame();
+            BoardDto dto = new BoardDto(game);
             printBoard(dto);
         }
     }
