--- src/main/java/chess/domain/square/Color.java
@@ -1,3 +1,0 @@
-package chess.domain.square;
-
-public enum Color { WHITE, BLACK }
