--- src/main/java/chess/domain/square/Empty.java
@@ -1,21 +1,0 @@
-package chess.domain.square;
-
-public class Empty implements Square {
-
-    private static final String DISPLAY = ".";
-
-    @Override
-    public String display() {
-        return DISPLAY;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        return o != null && getClass() == o.getClass();
-    }
-
-    @Override
-    public String toString() {
-        return "Empty";
-    }
-}
