--- src/main/java/chess/domain/square/Piece.java
@@ -1,10 +1,0 @@
-package chess.domain.square;
-
-public abstract class Piece implements Square {
-
-    protected final Color color;
-
-    protected Piece(Color color) {
-        this.color = color;
-    }
-}
