--- src/main/java/chess/domain/square/Square.java
@@ -1,6 +1,0 @@
-package chess.domain.square;
-
-public interface Square {
-
-    String display();
-}
