+++ src/test/java/chess/domain/piece/QueenTest.java
@@ -1,4 +1,4 @@
-package chess.piece;
+package chess.domain.piece;
 
 import static chess.domain.piece.Color.BLACK;
 import static chess.domain.piece.Color.WHITE;
