+++ src/test/java/chess/domain/piece/position/PositionTest.java
@@ -1,9 +1,8 @@
-package chess.piece.position;
+package chess.domain.piece.position;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-import chess.domain.piece.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
