+++ src/test/java/chess/domain/piece/position/PositionUtilTest.java
@@ -1,4 +1,4 @@
-package chess.piece.position;
+package chess.domain.piece.position;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
